fix: unblock PR CI checks
